pyproject.toml
[project]
name = "exobrain"
version = "0.2.0"
description = "A toy AI assistant with agent capabilities and file system access"
authors = [{ name = "VisualDust", email = "gavin@gong.host" }]
requires-python = ">=3.10,<4"
readme = "README.md"
license = { text = "MIT" }
keywords = [
    "ai",
    "assistant",
    "agent",
    "cli",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.5.0,<3",
    "pydantic-settings>=2.1.0,<3",
    "rich>=13.7.0,<14",
    "click>=8.1.7,<9",
    "aiofiles>=23.2.1,<24",
    "httpx>=0.26.0,<0.27",
    "pyyaml>=6.0.1,<7",
    "python-dotenv>=1.0.0,<2",
    "beautifulsoup4>=4.12.0,<5",
    "lxml>=5.0.0,<6",
    "pre-commit>=3.8.0",
    "textual>=0.89.0",
    "ddgs>=9.5.2",
    "pypdf>=4.0.0,<5",
]

[project.optional-dependencies]
openai = ["openai>=1.7.0,<2"]
anthropic = ["anthropic>=0.8.0,<0.9"]
all = [
    "openai>=1.7.0,<2",
    "anthropic>=0.8.0,<0.9",
]

[project.urls]
Homepage = "https://github.com/visualDust/exobrain"
Repository = "https://github.com/visualDust/exobrain"

[project.scripts]
exobrain = "exobrain.cli:main"
gavin = "exobrain.cli:main" # Alias for exobrain

[dependency-groups]
dev = [
    "pytest>=7.4.3,<8",
    "pytest-asyncio>=0.23.2,<0.24",
    "pytest-cov>=4.1.0,<5",
    "pytest-mock>=3.12.0,<4",
    "black>=23.12.1,<24",
    "ruff>=0.1.9,<0.2",
    "mypy>=1.8.0,<2",
    "pre-commit>=3.6.0,<4",
    "ipython>=8.19.0,<9",
]
docs = [
    "mkdocs>=1.5.3,<2",
    "mkdocs-material>=9.5.3,<10",
    "mkdocstrings[python]>=0.24.0,<0.25",
]

[tool.uv]
default-groups = [
    "dev",
    "docs",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "llama_cpp.*",
    "anthropic.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=exobrain --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["exobrain"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
